requirements.txt
python-pptx==1.0.2

# Python version
python_version >= "3.8"
